Task 1/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
pydantic==2.4.2
python-multipart==0.0.6
httpx==0.25.1
requests==2.31.0
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
python-magic==0.4.27
aiohttp==3.9.1
beautifulsoup4==4.12.2 